services/pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "wearforce-services"
version = "0.1.0"
description = "WearForce CRM/ERP Microservices"
authors = ["WearForce Team <dev@wearforce.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
sqlmodel = "^0.0.14"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
redis = "^5.0.1"
celery = "^5.3.4"
strawberry-graphql = {extras = ["fastapi"], version = "^0.214.0"}
nats-py = "^2.7.2"
pydantic = {extras = ["email"], version = "^2.5.0"}
pydantic-settings = "^2.1.0"
structlog = "^23.2.0"
httpx = "^0.25.2"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
gunicorn = "^21.2.0"
prometheus-client = "^0.19.0"
# Authentication & Security
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
# Email & SMS
aiosmtplib = "^3.0.1"
twilio = "^8.10.0"
# Push notifications
firebase-admin = "^6.4.0"
# Background tasks
rq = "^1.15.1"
# CLI tool
typer = "^0.9.0"
# Testing
pytest = {version = "^7.4.3", optional = true}
pytest-asyncio = {version = "^0.21.1", optional = true}
pytest-cov = {version = "^4.1.0", optional = true}
factory-boy = {version = "^3.3.0", optional = true}
faker = {version = "^20.1.0", optional = true}

[tool.poetry.extras]
test = ["pytest", "pytest-asyncio", "pytest-cov", "factory-boy", "faker"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100